cast needed on big-endian platforms
git-svn-id: https://svn.r-project.org/R/trunk@66952 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/modules/internet/internet.c
@@ -493,9 +493,9 @@ static SEXP in_do_download(SEXP args)
                 if (nbytes > 1024 * 1024)
                     REprintf("downloaded %0.1f MB\n\n", (double)nbytes / 1024 / 1024, url);
                 else if (nbytes > 10240)
-                    REprintf("downloaded %d KB\n\n", nbytes / 1024, url);
+                    REprintf("downloaded %d KB\n\n", (int)nbytes / 1024, url);
                 else
-                    REprintf("downloaded %d bytes\n\n", nbytes, url);
+                    REprintf("downloaded %d bytes\n\n", (int)nbytes, url);
             }
 #ifdef Win32
             R_FlushConsole();
@@ -624,9 +624,9 @@ static SEXP in_do_download(SEXP args)
                 if (nbytes > 1024 * 1024)
                     REprintf("downloaded %0.1f MB\n\n", (double)nbytes / 1024 / 1024, url);
                 else if (nbytes > 10240)
-                    REprintf("downloaded %d KB\n\n", nbytes / 1024, url);
+                    REprintf("downloaded %d KB\n\n", (int)nbytes / 1024, url);
                 else
-                    REprintf("downloaded %d bytes\n\n", nbytes, url);
+                    REprintf("downloaded %d bytes\n\n", (int)nbytes, url);
             }
 #ifdef Win32
             R_FlushConsole();
